+++ Shared/Data Models/Network/Models/RedditPost.swift
@@ -8,7 +8,7 @@ struct RedditPost: RedditResponsable, RedditIdentifiable {
 	let title: String
 	let author: String
 	let score: Int
-	let scoreProportion: Float
+	let scoreProportion: Double
 	let commentCount: Int
 	let createdAt: TimeInterval
 	let editedAt: TimeInterval?
@@ -37,7 +37,7 @@ struct RedditPost: RedditResponsable, RedditIdentifiable {
 //		print(json, hashID) //SAMPLE
 		author = data["author"] as! String
 		score = data["score"] as! Int
-		scoreProportion = data["upvote_ratio"] as! Float
+		scoreProportion = data["upvote_ratio"] as! Double
 		commentCount = data["num_comments"] as! Int
 		createdAt = data["created"] as! TimeInterval
 		let editTimestamp = data["edited"] as! TimeInterval
